requirements.txt
Django==1.6
Unipath==1.0
dj_database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
django-decouple==2.1
pytz
south==0.8.4
bootstrap-admin==0.3.0

